Added a lib taget to ravedude
Preliminary support for `cargo avr` tooling
