+++ ravedude/src/board/config.rs
@@ -1,0 +1,71 @@
+use crate::avrdude::config::BoardAvrdudeOptions;
+
+#[derive(serde::Serialize, serde::Deserialize, Debug, Default)]
+#[serde(rename_all = "kebab-case")]
+pub struct BoardConfig {
+    pub name: Option<String>,
+    pub inherit: Option<String>,
+    pub reset: Option<ResetOptions>,
+    pub avrdude: Option<BoardAvrdudeOptions>,
+    pub usb_info: Option<BoardUSBInfo>,
+}
+
+impl BoardConfig {
+    pub fn merge(self, base: BoardConfig) -> Self {
+        Self {
+            name: self.name.or(base.name),
+            // inherit is used to decide what BoardConfig to inherit and isn't used anywhere else
+            inherit: None,
+            reset: self.reset.or(base.reset),
+            avrdude: match self.avrdude {
+                Some(avrdude) => base.avrdude.map(|base_avrdude| avrdude.merge(base_avrdude)),
+                None => base.avrdude,
+            },
+            usb_info: self.usb_info.or(base.usb_info),
+        }
+    }
+}
+
+#[derive(serde::Serialize, serde::Deserialize, Debug)]
+pub struct ResetOptions {
+    pub automatic: bool,
+}
+
+#[derive(serde::Serialize, serde::Deserialize, Debug)]
+#[serde(rename_all = "kebab-case")]
+pub enum BoardUSBInfo {
+    PortIds(Vec<BoardPortID>),
+    #[serde(rename = "error")]
+    Error(String),
+}
+
+#[derive(serde::Serialize, serde::Deserialize, Debug)]
+pub struct BoardPortID {
+    pub vid: u16,
+    pub pid: u16,
+}
+
+impl BoardConfig {
+    pub fn guess_port(&self) -> Option<anyhow::Result<std::path::PathBuf>> {
+        match &self.usb_info {
+            Some(BoardUSBInfo::Error(err)) => Some(Err(anyhow::anyhow!(err.clone()))),
+            Some(BoardUSBInfo::PortIds(ports)) => {
+                for serialport::SerialPortInfo {
+                    port_name,
+                    port_type,
+                } in serialport::available_ports().unwrap()
+                {
+                    if let serialport::SerialPortType::UsbPort(usb_info) = port_type {
+                        for &BoardPortID { vid, pid } in ports {
+                            if usb_info.vid == vid && usb_info.pid == pid {
+                                return Some(Ok(port_name.into()));
+                            }
+                        }
+                    }
+                }
+                Some(Err(anyhow::anyhow!("Serial port not found.")))
+            }
+            None => None,
+        }
+    }
+}
