+++ ravedude/src/board/mod.rs
@@ -1,6 +1,6 @@
-use std::{collections::HashMap, path::Path};
+pub mod config;
 
-use crate::config;
+use std::collections::HashMap;
 
 fn get_all_boards() -> anyhow::Result<HashMap<String, config::BoardConfig>> {
     toml::from_str(include_str!("boards.toml")).map_err(|err| {
@@ -18,50 +18,20 @@ fn get_all_boards() -> anyhow::Result<HashMap<String, config::BoardConfig>> {
     })
 }
 
-pub fn get_board_from_name(board_name: &str) -> anyhow::Result<config::RavedudeConfig> {
+pub fn get_board_from_name(board_name: &str) -> anyhow::Result<config::BoardConfig> {
     let mut all_boards = get_all_boards()?;
 
-    Ok(config::RavedudeConfig {
-        board_config: Some(all_boards.remove(board_name).ok_or_else(|| {
-            let mut msg = format!("invalid board: {board_name}\n");
+    Ok(all_boards.remove(board_name).ok_or_else(|| {
+        let mut msg = format!("invalid board: {board_name}\n");
 
-            msg.push_str("valid boards:");
+        msg.push_str("valid boards:");
 
-            for board in all_boards.keys() {
-                msg.push('\n');
-                msg.push_str(&board);
-            }
-            anyhow::anyhow!(msg)
-        })?),
-        ..Default::default()
-    })
-}
-
-pub fn get_board_from_manifest(manifest_path: &Path) -> anyhow::Result<config::RavedudeConfig> {
-    Ok({
-        let file_contents = std::fs::read_to_string(manifest_path)
-            .map_err(|err| anyhow::anyhow!("Ravedude.toml read error:\n{}", err))?;
-
-        let mut board: config::RavedudeConfig = toml::from_str(&file_contents)
-            .map_err(|err| anyhow::anyhow!("invalid Ravedude.toml:\n{}", err))?;
-
-        if let Some(board_config) = board.board_config.as_ref() {
-            if let Some(board_name) = board.general_options.board.as_deref() {
-                anyhow::bail!(
-                    "can't both have board in [general] and [board] section; set inherit = \"{}\" under [board] to inherit its options",
-                    board_name
-                )
-            }
-            if let Some(inherit) = board_config.inherit.as_deref() {
-                let base_board = get_board_from_name(inherit)?.board_config.unwrap();
-                board.board_config = Some(board.board_config.take().unwrap().merge(base_board));
-            }
-        } else if let Some(board_name) = board.general_options.board.as_deref() {
-            let base_board = get_board_from_name(board_name)?.board_config.unwrap();
-            board.board_config = Some(base_board);
+        for board in all_boards.keys() {
+            msg.push('\n');
+            msg.push_str(&board);
         }
-        board
-    })
+        anyhow::anyhow!(msg)
+    })?)
 }
 
 #[cfg(test)]
